requirements.txt
langchain-community>=0.0.10
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
numpy<2
docarray>=0.30.0
chromadb>=0.4.0
yandexcloud